Step Selector: Persist step selection settings. (#5975)
* Motivation for features / changes
The new step selection feature currently must be enabled every time you
load Tensorboard. This is annoying for users who like this feature and
have to turn it on constantly. This change persists the settings in
LocalStorage so it can stay enabled on that machine until it is
disabled.

* Technical description of changes
Just followed the pattern of settings that are already persisted.