linked time: update scalar to "real" min max step (#5993)
* Motivation for features / changes
The "minMax" input of scalar fob controller is actually the
viewExtend/viewBox, not the min max step, while it's got used as the min
max step. Visually the extend is larger than the min/max step and it is
what we need for forward and reverse function.
```
            A   B   (chart)   C   D
            |---|-------------|---|
```
axis size: A to D
minMax:  A to D
line chart area: B to C
min step: B
max step C

This wasn't noticed because we clip the time selection at the scalar
card level to prevent the fob go out of the valid range (B to C)

* Technical description of changes
This PR adds another minMaxStep input and rename the old minMax to
minMaxHorizentalViewExtend.
(Note: skip the rename in test since that matters less)

* Screenshots of UI changes
Everything should be the same.