Data Table: Use smoothed values in table (#6042)
* Motivation for features / changes
When users have smoothing enabled the raw values on the scalar chart are
faded and hidden in the background while the smoothed lines are
highlighted and in the foreground. However, while in range selection
mode the data table only showed values based on the real values. This
was confusing for users. This changes the data table to use the smoothed
and highlighted values on the chart in the data table.

* Technical description of changes
We simply changed the calculations to use the "y" property on the
ScalarCardPoint object instead of the "value" property.