Column Customization: Move header order to NGRX (#6078)
* Motivation for features / changes
This is part of the Column Customization effort which will allow users
to select which headers they want to appear in the scalar card data
table and in which order. In order to do this we need to have that
selection and order stored in the ngrx state . This change simply moves
that information from being set in the ScalarCardContainer to being set
in the default ngrx state.

* Technical description of changes
In order to make no visible changes in production I added special logic
to remove the smoothed column if smoothing is disabled. I think there
should be some discussion on how this is handled in the future.

* Screenshots of UI changes
No visible change to users