Upgrade to TypeScript 4.8.4. (#6082)
Upgrade to TypeScript 4.8.4 following instructions in
https://github.com/tensorflow/tensorboard/blob/master/DEVELOPMENT.md#adding-updating-or-removing-frontend-dependencies.

This upgrade is also a good opportunity to remove some @ts-ignore
annotations that were added to the codebase to satisfy the internal
upgrade to TS 4.8.4. See
https://github.com/tensorflow/tensorboard/pull/6075 and
https://github.com/tensorflow/tensorboard/pull/6037. Removing the
annotations did indeed lead to build errors that I then fixed with some
type adjustments. There is no change in behavior.

Verified by:
* Running the Angular test suite.
* Building and running the TensorBoard app and clicking around.
* `cd tensorboard; grep -r "ts-ignore" *` and ensuring all relevant
   @ts-ignore annotations have been removed.