Use `gcp_auth` for GCP authentication (#5939)
## Motivation for features / changes

See #5934 

At the moment, there are 2 types of `Credentials`:

- `Anonymous` (default),
- `RefreshToken` (effectively, loaded from a file with credentials).

In GKE service managed account, none of those would work if we want to
access non-public data. Accessing public data works via the `Anonymous`
access. We cannot use file-based credentials management for security
reasons.

This strongly limits the usage of the `tensorboard --load-fast=true`
application on GKE / k8s deployed instances. Enabling the `rustboard`
backend to be compatible with GKE service managed accounts will
positively affect the adoption and operations of the `tensorboard` tool.

## Technical description of changes

In this PR, the existing code for GCP authentication (via `Credentials`)
is replaced with `gcp_auth`. This requires a new dependency:
[`gcp_auth`](https://docs.rs/gcp_auth/0.7.3/gcp_auth/). That cargo
manages pretty much everything related to GCP authentication

##  Screenshots of UI changes

None

## Detailed steps to verify changes work correctly (as executed by you)

We built the `rustboard` from the last commit. We replaced the existing
executable for `rustboard` that comes as part of the `tensorboard`
package with the custom build. We executed the `tensorboard` command
from the GKE machine and confirmed that the access is granted as
expected (service account authentication was successful).

## Alternate designs / implementations considered

Happy to hear about that.